move texture image into it's own module
because:
- it's fairly stand-alone functionality
